requirements.txt
google-genai
langchain-google-genai
langchain
python-dotenv